feat: refactor chat-related functions and constants into chatHelpers utility file
